% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bullet-scores.R
\name{bootstrap_k}
\alias{bootstrap_k}
\title{Helper function: bootstrap scores}
\usage{
bootstrap_k(scores, k, K, value)
}
\arguments{
\item{scores}{numeric values of scores}

\item{k}{number of lands to average across}

\item{K}{number of replicates}

\item{value}{observed value}
}
\description{
Helper function: bootstrap scores
}
