% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{bullet_pipeline}
\alias{bullet_pipeline}
\title{Helper file to setup data}
\usage{
bullet_pipeline(location, stop_at_step = NULL, x3p_clean = function(x) x, ...)
}
\arguments{
\item{location}{directory or list of urls containing x3p files.
If there are sub-directories or sub-lists, this function will assume
that each sub-directory contains lands from different bullets and
will separate them accordingly}

\item{stop_at_step}{One of read, clean, crosscut, grooves, signatures}

\item{x3p_clean}{function to use to clean the x3p file - convert header info
to correct units, rotate the surface matrix, etc.}

\item{...}{additional arguments to cc_locate_grooves and cc_get_signature}
}
\value{
a data_frame with bullet x3p files and processed data steps
}
\description{
Helper file to setup data
}
