% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bullet-scores.R
\name{bullet_to_land_predict}
\alias{bullet_to_land_predict}
\title{Get land to land prediction based on bullet to bullet comparisons}
\usage{
bullet_to_land_predict(land1, land2, scores)
}
\arguments{
\item{land1}{(numeric) vector with land ids of bullet 1}

\item{land2}{(numeric) vector with land ids of bullet 2}

\item{score}{numeric vector of scores to be summarized into a single number}
}
\value{
numeric vector of binary prediction whether two lands are same-source. Vector has the same length as the input vectors.
}
\description{
The combination of `land1` and `land2` are a key to the scores,
i.e. if a bullet has six lands, each of the input vectors should have
length 36.
}
