% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loess.R
\name{cc_fit_loess}
\alias{cc_fit_loess}
\title{Fit a loess curve to a bullet data frame}
\usage{
cc_fit_loess(ccdata, span = 0.75)
}
\arguments{
\item{ccdata}{The crosscut as returned from x3p_to_df}

\item{span}{The span to use for the loess regression}
}
\value{
a list of a data frame of the original bullet measurements extended by loess fit, residuals, and standard errors and two plots: a plot of the fit, and a plot of the bullet's land signature.
}
\description{
A loess regression is fit to the surface measurements and residuals are calculated.
The most extreme 0.25% of residuals are filtered from further consideration.
The result is called the raw signature of the bullet land.
}
