% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signatures.R
\name{cc_get_signature}
\alias{cc_get_signature}
\title{Extract signature from crosscut}
\usage{
cc_get_signature(ccdata, grooves = NULL, span1 = 0.75, span2 = 0.03)
}
\arguments{
\item{ccdata}{crosscut as returned from x3p_crosscut}

\item{grooves}{The grooves to use as a two element vector, if desired}

\item{span1}{The span for the loess fit to get from the profile to the raw signature}

\item{span2}{The span for the loess fit to smooth the raw signature}
}
\value{
data frame
}
\description{
x3p file of a 3d topological bullet surface is processed at surface
crosscut y,
measurements outside the bullet grooves in the crosscuts are left out, and
a loess smooth
is used (see \code{?loess} for details) to remove the big structure.
}
\examples{
data(br411)
br411_processed <- processBullets(br411, name = "br411")
}
