% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signatures.R
\name{cc_get_signature_notused}
\alias{cc_get_signature_notused}
\title{Smooth the raw signature to remove excess variability}
\usage{
cc_get_signature_notused(data, span = 0.03, limits = c(-5, 5),
  id = "bullet")
}
\arguments{
\item{data}{data frame as returned by the function \code{cc_get_raw_signature}}

\item{span}{width of the smoother, defaults to 0.03}

\item{limits}{vector of the form c(min, max). Results will be limited to be between these values.}

\item{id}{variable name of the land identifier}
}
\value{
data frame of the same form as the input extended by the vector sig for the smoothed signature.
}
\description{
Smooth the raw signature to remove excess variability
}
