% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crosscut.R
\name{check_ccdata}
\alias{check_ccdata}
\title{Check object returned by x3p_crosscut_optimize}
\usage{
check_ccdata(x)
}
\arguments{
\item{x}{data frame from x3p_crosscut_optimize}
}
\value{
TRUE if everything is ok, error otherwise
}
\description{
Check object returned by x3p_crosscut_optimize
}
