% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loess.R
\name{check_loess_fit}
\alias{check_loess_fit}
\title{Check loess or gaussian curve fit object}
\usage{
check_loess_fit(x)
}
\arguments{
\item{x}{output from cc_fit_loess or cc_fit_gaussian}
}
\value{
TRUE or error
}
\description{
Check loess or gaussian curve fit object
}
