% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peaks.R
\name{check_striae}
\alias{check_striae}
\title{Check a striae object is valid}
\usage{
check_striae(striae)
}
\arguments{
\item{striae}{data frame with a striae object}
}
\value{
TRUE if all conditions are met, error otherwise
}
\description{
Check a striae object is valid
}
