% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crosscut.R
\name{check_x3p}
\alias{check_x3p}
\title{Check whether an x3p argument is character or filename, return an x3p object}
\usage{
check_x3p(x3p)
}
\arguments{
\item{x3p}{if character, path to an x3p file. Otherwise a scan in x3p
format is expected.}
}
\description{
Check whether an x3p argument is character or filename, return an x3p object
}
