% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bullet-scores.R
\name{compute_average_scores}
\alias{compute_average_scores}
\title{Get average scores for bullet to bullet comparisons}
\usage{
compute_average_scores(land1, land2, score)
}
\arguments{
\item{land1}{(numeric) vector with land ids of bullet 1}

\item{land2}{(numeric) vector with land ids of bullet 2}

\item{score}{numeric vector of scores to be summarized into a single number}
}
\value{
numeric vector of average scores. Length is the same as the number of
          land engraved areas on the bullets.
}
\description{
Note that the combination of `land1` and `land2` are a key to the scores,
i.e. if a bullet has six lands, each of the input vectors should have
length 36.
}
