% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c_wrapper.R
\name{compute_cross_corr}
\alias{compute_cross_corr}
\title{Wrapper function for compute_cross_corr}
\usage{
compute_cross_corr(x, y, min.overlap)
}
\arguments{
\item{x}{numeric vector, the longer sequence}

\item{y}{numeric vector, the shorter sequence}

\item{min.overlap}{numeric scalar, set the length of the minimum overlapping part}
}
\description{
Wrapper function for compute_cross_corr
}
