% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{extract_feature_ccf}
\alias{extract_feature_ccf}
\title{Extract ccf from two (or more) aligned signatures}
\usage{
extract_feature_ccf(aligned)
}
\arguments{
\item{aligned}{data frame with variable x (for location) and two or
more measurements}
}
\value{
(matrix) of correlations
}
\description{
Extract ccf from two (or more) aligned signatures
}
