% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{extract_feature_cms2}
\alias{extract_feature_cms2}
\title{Extract number of consecutively matching elevated striation marks from two aligned signatures}
\usage{
extract_feature_cms2(striae)
}
\arguments{
\item{striae}{data frame of striation marks based on two aligned signatures}
}
\value{
number of consecutively matching elevated striation marks
}
\description{
Extract number of consecutively matching elevated striation marks from two aligned signatures
}
