% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{extract_feature_cms_per_mm}
\alias{extract_feature_cms_per_mm}
\title{Extract scaled number of consecutively matching striation marks (peaks and valleys) from two aligned signatures}
\usage{
extract_feature_cms_per_mm(striae, aligned, resolution)
}
\arguments{
\item{striae}{data frame of striation marks based on two aligned signatures}

\item{aligned}{two aligned signatures}

\item{resolution}{micron per pixel resolution}
}
\value{
scaled number of consecutively matching striation marks (peaks and valleys)
}
\description{
Extract scaled number of consecutively matching striation marks (peaks and valleys) from two aligned signatures
}
\examples{
\dontrun{
# Set the data up to be read in, cleaned, etc.
library(bulletxtrctr)
library(x3ptools)

example_data <- bullet_pipeline(
  location = list(
    Bullet1 = c(hamby252demo$bullet1[2]),
    Bullet2 = c(hamby252demo$bullet2[4])
  ),
  x3p_clean = function(x) x \%>\%
      x3p_scale_unit(scale_by=10^6) \%>\%
      rotate_x3p(angle = -90) \%>\%
      y_flip_x3p()
)

alignment <- sig_align(example_data$sigs[[1]]$sig,
                       example_data$sigs[[2]]$sig)
striae <- sig_cms_max(alignment)
extract_feature_cms(striae$lines)
}
}
\concept{scaled-features}
