% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{extract_feature_D}
\alias{extract_feature_D}
\title{Extract average distance between two (or more) aligned signatures}
\usage{
extract_feature_D(aligned, ...)
}
\arguments{
\item{aligned}{data frame with variable x (for location) and two or more measurements}

\item{...}{arguments for function `dist`}
}
\value{
object of class distance
}
\description{
Extract average distance between two (or more) aligned signatures
}
