% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{extract_feature_lag_mm}
\alias{extract_feature_lag_mm}
\title{Extract lag in mm from two (or more) aligned signatures}
\usage{
extract_feature_lag_mm(aligned, resolution)
}
\arguments{
\item{aligned}{data frame with variable x (for location) and two or
more measurements (the bullets entry in the list returned from
sig_align)}

\item{resolution}{numeric value of resolution in micron per pixel}
}
\value{
(vector) of lags in millimeter
}
\description{
Extract lag in mm from two (or more) aligned signatures
}
\seealso{
Other alignment-related-features: 
\code{\link{extract_feature_D}()},
\code{\link{extract_feature_ccf}()},
\code{\link{extract_feature_lag}()},
\code{\link{extract_feature_length_mm}()},
\code{\link{extract_feature_length}()},
\code{\link{extract_feature_overlap}()},
\code{\link{extract_features_all}()}
}
\concept{alignment-related-features}
