% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{extract_feature_left_cms}
\alias{extract_feature_left_cms}
\title{Extract number of consecutively matching elevated striation marks from the
left of two aligned signatures}
\usage{
extract_feature_left_cms(striae)
}
\arguments{
\item{striae}{data frame of striation marks based on two aligned signatures}
}
\value{
number of consecutively matching striation marks (from left)
}
\description{
Extract number of consecutively matching elevated striation marks from the
left of two aligned signatures
}
\examples{
\dontrun{
# Set the data up to be read in, cleaned, etc.
library(bulletxtrctr)
library(x3ptools)

example_data <- bullet_pipeline(
  location = list(
    Bullet1 = c(hamby252demo$bullet1[2]),
    Bullet2 = c(hamby252demo$bullet1[4])
  ),
  x3p_clean = function(x) x \%>\%
      x3pheader_to_microns() \%>\%
      rotate_x3p(angle = -90) \%>\%
      y_flip_x3p()
)

alignment <- sig_align(example_data$sigs[[1]]$sig,
                       example_data$sigs[[2]]$sig)
striae <- sig_cms_max(alignment)
extract_feature_left_cms(striae$lines)
}
}
