% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{extract_feature_length}
\alias{extract_feature_length}
\title{Extract length of two (aligned) signatures}
\usage{
extract_feature_length(aligned)
}
\arguments{
\item{aligned}{data frame with variable x (for location) and two or more measurements}
}
\value{
integer value of the length of the shorter signature.
}
\description{
Signatures will usually be of different lengths. In a comparison, the length of the shorter signature represents the potential length for a match.
}
