% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{extract_feature_matches}
\alias{extract_feature_matches}
\title{Extract number of matching striation marks from two aligned signatures}
\usage{
extract_feature_matches(striae)
}
\arguments{
\item{striae}{data frame of striation marks based on two aligned signatures}
}
\value{
number of matching striation marks
}
\description{
Extract number of matching striation marks from two aligned signatures
}
