% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{extract_feature_mismatches}
\alias{extract_feature_mismatches}
\title{Extract number of mismatched striation marks from two aligned signatures}
\usage{
extract_feature_mismatches(striae)
}
\arguments{
\item{striae}{data frames of striation marks based on two aligned signatures}
}
\value{
number of mismatched striation marks
}
\description{
Extract number of mismatched striation marks from two aligned signatures
}
