% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{extract_feature_n_striae}
\alias{extract_feature_n_striae}
\title{Extract information for striation marks from two aligned signatures}
\usage{
extract_feature_n_striae(striae, type = "peak", match = TRUE)
}
\arguments{
\item{striae}{data frames of striation marks based on two aligned signatures}

\item{type}{one of "peak", "valley" or "all"}

\item{match}{binary setting: TRUE for matching striae, FALSE for non-matching
striae}
}
\value{
number of striae
}
\description{
internal function, called by multiple extract_feature functions
}
