% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{extract_feature_overlap}
\alias{extract_feature_overlap}
\title{Extract overlap between two aligned signatures}
\usage{
extract_feature_overlap(aligned)
}
\arguments{
\item{aligned}{data frame with variable x (for location) and two or more measurements}
}
\value{
value between 0 and 1, ratio of length of overlap compared to smaller length of the signature
}
\description{
The overlap of two aligned signatures is defined as the ratio of the number of non-missing overlapping values of the two aligned signatures and
the length of the shorter signature. A larger overlap indicates a higher level of agreement between the signatures.
}
