% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{extract_feature_sd_D}
\alias{extract_feature_sd_D}
\title{Extract variation in the height measurements between two aligned signatures}
\usage{
extract_feature_sd_D(aligned)
}
\arguments{
\item{aligned}{data frame with variable x (for location) and two aligned
signatures}
}
\value{
numeric value
}
\description{
Extract variation in the height measurements between two aligned signatures
}
