% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{extract_feature_sum_peaks}
\alias{extract_feature_sum_peaks}
\title{Extract the combined height of aligned striae between two aligned signatures}
\usage{
extract_feature_sum_peaks(striae)
}
\arguments{
\item{striae}{data frame of striation marks based on two aligned signatures}
}
\value{
sum of heights of matching striae
}
\description{
Extract the combined height of aligned striae between two aligned signatures
}
