% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{extract_helper_feature_n_striae}
\alias{extract_helper_feature_n_striae}
\title{Extract information for striation marks from two aligned signatures}
\usage{
extract_helper_feature_n_striae(striae, type = "peak", match = TRUE)
}
\arguments{
\item{striae}{data frame of striation marks based on two aligned signatures}

\item{type}{one of "peak", "valley" or "all"}

\item{match}{binary setting: TRUE for matching striae, FALSE for non-matching
striae}
}
\value{
number of striae
}
\description{
internal function, called by multiple extract_feature functions
}
\examples{
\dontrun{
# Set the data up to be read in, cleaned, etc.
library(bulletxtrctr)
library(x3ptools)

example_data <- bullet_pipeline(
  location = list(
    Bullet1 = c(hamby252demo$bullet1[2]),
    Bullet2 = c(hamby252demo$bullet2[4])
  ),
  x3p_clean = function(x) x \%>\%
      x3pheader_to_microns() \%>\%
      rotate_x3p(angle = -90) \%>\%
      y_flip_x3p()
)

alignment <- sig_align(example_data$sigs[[1]]$sig,
                       example_data$sigs[[2]]$sig)
striae <- sig_cms_max(alignment)
extract_helper_feature_n_striae(striae$lines, type = "all", match = T)
}
}
\seealso{
Other striae-related-features: \code{\link{extract_feature_cms_all}},
  \code{\link{extract_feature_cms_peak_only}},
  \code{\link{extract_feature_left_cms}},
  \code{\link{extract_feature_matches}},
  \code{\link{extract_feature_mismatches}},
  \code{\link{extract_feature_non_cms}},
  \code{\link{extract_feature_right_cms}},
  \code{\link{extract_feature_sum_peaks}},
  \code{\link{extract_features_all}}
}
