% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phase_test.R
\name{F_T}
\alias{F_T}
\alias{f_T}
\title{#' Extract results from test in tidy form
#'
#' \code{tidy.phase.test} expands the tidy method for test.phase objects. It gives a summary of the relevant
#' parameters and estimates.
#' @param x phase.test object as returned from \code{phase_test}
#' @param ... ignored
#' @export
#' @importFrom broom tidy
#' @examples
#' logo <- x3ptools::x3p_read(system.file("csafe-logo.x3p", package="x3ptools"))
#' print(logo)
tidy <- function (x, ...) {
with(x, tibble(estimate, estimate1, estimate2, statistic, p.value, parameter))
}
Reference distribution and debsity for the test statistic between Same-source and Different-source averages using phase selection}
\usage{
F_T(t, sigma = 1, n = 6, lower.tail = TRUE)

f_T(t, sigma = 1, n = 6)
}
\arguments{
\item{t}{value of the observed test statistic, t has to be non-negative.}

\item{sigma}{(estimated) standard deviation of an average of size n of the (original) sample}

\item{n}{integer value, number of values in the same-source group.}

\item{lower.tail}{should the lower tail probability be returned?}
}
\description{
Assuming an overall sample size of 36 (for 6 x 6 comparisons, and n=6 independent objects), we are interested in the distribution of
T defined as the difference between the average of the same source values and the median average (XXX see paper for definition) of the different source values:
\eqn{T = \bar{X}_{SS} - \tilde{X}_{DS}}, where groups SS and DS are set by phase selection (after inspecting the set of n values).
For a (one-sided) hypothesis test of \eqn{H_0}: no difference between the means and \eqn{H_1:}: the same-source values are larger
on average than the different source values, the test statistic T has under the null hypothesis and the group-finding
strategy \eqn{\cal G} a distribution given as
\deqn{ P(T < t \mid H_0, {\cal G}) = C \cdot \int_{-\infty}^{\infty} F^2(x) f(x) \left[ F(x + t) - F(x) \right]^3 dx}
where F is the (approximately normal) distribution of \eqn{\bar{X}_{SS}} under \eqn{H_0} with expected value of \eqn{\mu} (\eqn{=E[\bar{X}_{SS}] = E[\bar{X}_{DS}]}) and standard deviation \eqn{\sigma} (standard deviation of a group of size $n$ under $H_0$).
}
\examples{
# example code
F_T(0.2, sigma = 0.1, lower.tail = FALSE) # only multitude of sigma informs the test
F_T(2, lower.tail = FALSE)
F_T(3, lower.tail = FALSE)
F_T(4, lower.tail = FALSE)
# critical values of t distribution and F_T (for sigma = 1):
F_T(qt(1-0.05, df=34), lower.tail = FALSE)
F_T(qt(1-0.05/6, df=34), lower.tail = FALSE)
F_T(qt(1-0.01/6, df=34), lower.tail = FALSE)

x <- 10^-(1:4)
x <- sort(c(x, x/2))
t_q <- qt(1-x, df=34)
adjusted <- F_T(t_q, lower.tail=FALSE)
}
