% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grooves.R
\name{get_grooves_lassobasic}
\alias{get_grooves_lassobasic}
\title{Use logistic model to identify groove locations}
\usage{
get_grooves_lassobasic(x, value, pred_cutoff = 0.07, return_plot = F)
}
\arguments{
\item{x}{numeric vector of locations (in microns)}

\item{value}{numeric values of surface measurements in microns}

\item{pred_cutoff}{equal error rate cutoff for classification into GEA or LEA, trained on Hamby set 44}

\item{return_plot}{return plot of grooves?}
}
\description{
Use logistic model to identify groove locations
}
