% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grooves.R
\name{get_grooves_logistic}
\alias{get_grooves_logistic}
\title{Use logistic model to identify groove locations}
\usage{
get_grooves_logistic(x, value, smoothfactor = 15, adjust = 10,
  groove_cutoff = 400, return_plot = F)
}
\description{
Use logistic model to identify groove locations
}
