% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grooves.R
\name{get_grooves_middle}
\alias{get_grooves_middle}
\title{Use the center of a crosscut}
\usage{
get_grooves_middle(x, value, middle = 75)
}
\arguments{
\item{x}{numeric vector of locations in microns}

\item{value}{numeric vector of surface measurements in microns}

\item{middle}{middle percent to use for the identification}
}
\value{
list of groove vector and plot of crosscut with
}
\description{
Use the center of a crosscut
}
