% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grooves.R
\name{get_grooves_quadratic}
\alias{get_grooves_quadratic}
\title{Quadratic fit to find groove locations}
\usage{
get_grooves_quadratic(x, value, adjust)
}
\arguments{
\item{x}{numeric vector of locations (in microns)}

\item{value}{numeric values of surface measurements in microns}

\item{adjust}{positive number to adjust the grooves}
}
\value{
list of groove vector and plot of crosscut with shoulder locations
}
\description{
Use a robust fit of a quadratic curve to find groove locations
}
