% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cms.R
\name{get_longest_run}
\alias{get_longest_run}
\title{Number of maximum consecutively matching striae}
\usage{
get_longest_run(match)
}
\arguments{
\item{match}{is a Boolean vector of matches/non-matches}
}
\value{
an integer value of the maximum number of consecutive matches
}
\description{
Number of maximum consecutively matching striae
}
\examples{
x <- rbinom(100, size = 1, prob = 1/3)
get_run(x == 1) # expected value for longest match is 3
get_longest_run(x==1)
}
