% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bullet-scores.R
\name{get_phases}
\alias{get_phases}
\title{Get bullet phases}
\usage{
get_phases(land1, land2)
}
\arguments{
\item{land1}{vector with land ids of bullet 1}

\item{land2}{vector with land ids of bullet 2}
}
\value{
numeric vector of the phases (diagonal groupings).
}
\description{
Note that the combination of \code{land1} and \code{land2} are a key to the land-to-land scores,
i.e. if a bullet has six lands, each of the input vectors should have 36 entries. In the case that
the lands are not factor variables, the order in which different lands are encountered in the vector
is used to determine their order.
}
