% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cms.R
\name{get_run}
\alias{get_run}
\title{Table of the number of consecutive matches}
\usage{
get_run(match)
}
\arguments{
\item{match}{is a Boolean vector of matches/non-matches}
}
\value{
a table of the number of the CMS and their frequencies
}
\description{
Table of the number of consecutive matches
}
\examples{
x <- rbinom(100, size = 1, prob = 1/3)
get_runs(x == 1) # expected value for longest match is 3
}
