% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cms.R
\name{get_runs}
\alias{get_runs}
\title{Table of the number of runs}
\usage{
get_runs(x)
}
\arguments{
\item{x}{Boolean vector}
}
\value{
a table of the number of runs of TRUEs
}
\description{
Identify the length of runs (of values TRUE) and their frequencies.
}
\examples{
x <- rbinom(100, size = 1, prob = 1/3)
get_runs(x == 1) # expected value for longest run is 3
get_runs(x == 0) # expected value for longest run is 6
}
