% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sig_align_set.R
\name{get_sig_lags}
\alias{get_sig_lags}
\title{Get lags between signatures}
\usage{
get_sig_lags(
  data,
  value,
  sig_id,
  min.overlap = 500,
  align_to = levels(sig_id)[1],
  resolution = 0.645,
  tol = 1e-06
)
}
\arguments{
\item{data}{data frame with ids of signatures, x variable and \code{value} variable.}

\item{value}{symbol for the signature values}

\item{sig_id}{factor variable with signatures ids}

\item{min.overlap}{value passed on to \code{?get_ccf}.}

\item{align_to}{id of the signature used for alignment. Defaults to the first factor level.}

\item{resolution}{numeric value representing the resolution of the signature in microns per pixel.}

\item{tol}{tolerance used, i.e. smallest value, such that 1 + tol not equal to 1.
Can't be smaller than the machine's numeric resolution \code{.Machine$double.eps}.}
}
\value{
data frame with a summary (one row for each level of \code{sig_id}) containing
the information on \code{id}, \code{lag}, and \code{ccf}.
}
\description{
Extract lags for aligning all signatures in a set with respect to the signature specified in 'align_to'.
Alignments will vary depending on which signature is used for alignment.
}
\examples{
\dontrun{
tmaRks::data("toolmarks", package="tmaRks")
tnest <- toolmarks \%>\% filter(size=="L") \%>\% group_by(side, angle) \%>\% tidyr::nest()
tnest <- tnest \%>\% mutate(
  lags = data \%>\% purrr::map(.f = get_sig_lags, value=signature,
              sig_id=TID, min.overlap=400, resolution = 7.0505e-03)
)
}
}
