% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hamby252demo}
\alias{hamby252demo}
\title{hamby252demo}
\format{a list containing lists of links for 6 lands of each bullet}
\usage{
hamby252demo
}
\description{
Links to barrel 1, bullets 1 and 2, with all 6 lands.
}
\keyword{datasets}
