% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hamby252demo_github}
\alias{hamby252demo_github}
\title{hamby252demo_github}
\format{
a list containing lists of links for the 6 lands of bullet 1 and 2
}
\usage{
hamby252demo_github
}
\description{
Links to barrel 1, bullets 1 and 2, with all 6 lands, mirrored on github in case NBTRD is down
}
\keyword{datasets}
