% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crosscut.R
\name{land_cc}
\alias{land_cc}
\title{Get a specified cross section}
\usage{
land_cc(y, land)
}
\arguments{
\item{y}{y coordinate of cross section}

\item{land}{x3p file with land data}
}
\value{
cross cut data frame
}
\description{
Get a specified cross section
}
