% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bullet-scores.R
\name{max_u}
\alias{max_u}
\title{Wilcox test of bullet to bullet similarity}
\usage{
max_u(land1, land2, scores, addNA = FALSE)
}
\arguments{
\item{land1}{(numeric) vector with land ids of bullet 1}

\item{land2}{(numeric) vector with land ids of bullet 2}

\item{scores}{numeric vector of scores to be summarized into a single number}

\item{addNA}{how are missing values treated? addNA = TRUE leaves missing values, addNA=FALSE imputes with 0.}
}
\value{
numeric vector of binary prediction whether two lands are same-source. Vector has the same length as the input vectors.
}
\description{
The combination of \code{land1} and \code{land2} are a key to the scores,
i.e. if a bullet has six lands, each of the input vectors should have
length 36.
}
