% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phase_test.R
\name{print.phase.test}
\alias{print.phase.test}
\title{Print information of a phase test}
\usage{
\method{print}{phase.test}(x, ...)
}
\arguments{
\item{x}{phase.test object}

\item{...}{ignored}
}
\description{
\code{print.phase.test} expands the generic print method for x3p objects. It gives a summary of the most relevant x3p meta information and returns the object invisibly.
}
\examples{
logo <- x3ptools::x3p_read(system.file("csafe-logo.x3p", package="x3ptools"))
print(logo)
}
