% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_bullet}
\alias{read_bullet}
\title{Reading all scans from a folder}
\usage{
read_bullet(folder, ext = "x3p")
}
\arguments{
\item{folder}{character describing the path to a folder}

\item{ext}{character value, consisting of the extension(s) describing the file format the scans are in}
}
\value{
data frame with two variables, source and x3p, containing the path to the file and the corresponding x3p file
}
\description{
some more description
}
\examples{
 b1 <- read_bullet("data/Bullet1", "x3p")
 b2 <- read_bullet("data/Bullet2", "x3p")
}
