% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_bullet.R
\name{read_dir}
\alias{read_dir}
\title{Reading all x3p scans belonging to a folder}
\usage{
read_dir(path, extension = "x3p", hierarchy = c("land", "bullet",
  "barrel", "set"))
}
\arguments{
\item{path}{character describing the path to a folder}

\item{extension}{character value, consisting of the extension(s) describing the
file format the scans are in}

\item{hierarchy}{vector of characters describing the folder structure, starting from lowest level to highest.}
}
\value{
data frame of x3p files with appropriate meta information
}
\description{
Read all scans of a specified format from a folder. This operation is recursive, i.e. also reads scans from the folder of a folder.
}
