% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grooves.R
\name{resid_calc}
\alias{resid_calc}
\title{Calculate predicted values and residuals}
\usage{
resid_calc(cc, fit)
}
\arguments{
\item{cc}{crosscut data, including columns value_std and x}

\item{fit}{fitted loess object from robust_loess_fit}
}
\description{
Calculate predicted values and residuals
}
