% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rtrees}
\alias{rtrees}
\title{randomforest}
\format{
a random forest object fitted by the randomforest function from the package of the same name
}
\usage{
rtrees
}
\description{
This randomforest was fitted to predict known matches and non-matches from the scans of land engraved areas of the Hamby study (see Hare et al ).
}
\keyword{datasets}
