% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runExample.R
\name{runExample}
\alias{runExample}
\title{Execute the Shiny App for bullet investigation}
\usage{
runExample()
}
\description{
This shiny app would expect a tibble object called \code{shiny.tt} in the user's working environment.
}
\details{
shiny.tt should have variables: \code{x3p}, \code{scan_id}, \code{crosscut}, \code{ccdata}, \code{grooves}.

To have this shiny app run properly, \code{shiny.tt} should at least include \code{grooves} as a variable
}
\examples{
\dontrun{
bulletxtrctr::runExample()
}
}
