% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align.R
\name{sig_align}
\alias{sig_align}
\title{Align two surface cross cuts according to maximal correlation}
\usage{
sig_align(sig1, sig2)
}
\arguments{
\item{sig1}{vector of first signature}

\item{sig2}{vector of second signature}
}
\value{
list consisting of a) the maximal cross correlation, b) the lag resulting in the highest cross correlation, and c) same data frame as input, but y vectors are aligned for maximal correlation
}
\description{
The bullet with the first name serves as a reference, the second bullet is shifted.
Function copied from `bulletAlign`
}
