% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peaks.R
\name{sig_get_peaks}
\alias{sig_get_peaks}
\title{Identify the location and the depth of peaks and valleys in a signature}
\usage{
sig_get_peaks(sig, smoothfactor = 35, striae = TRUE, window = TRUE)
}
\arguments{
\item{sig}{numeric vector of smoothed signature, padded with NAs at front
and/or back.}

\item{smoothfactor}{set to default of 35. Smaller values will pick up on
smaller changes in the crosscut.}

\item{striae}{If TRUE, show the detected striae on the plot}

\item{window}{If TRUE, show the window of the striae on the plot}
}
\value{
list of several objects:
}
\description{
Adapted from the function `get_peaks`
}
