% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth.R
\name{smoothloess}
\alias{smoothloess}
\title{Predict smooth from a fit}
\usage{
smoothloess(y, span)
}
\arguments{
\item{y}{numeric vector}

\item{span}{The span of the loess fit}
}
\description{
Predict smooth from a fit
}
