% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crosscut.R
\name{switch_xy}
\alias{switch_xy}
\title{Switch x and y in a data frame}
\usage{
switch_xy(dataframe)
}
\arguments{
\item{dataframe}{a data frame with columns x and y}
}
\value{
dataframe
}
\description{
Used to relable x3ps
}
