% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phase_test.R
\name{tidy}
\alias{tidy}
\title{Extract results from test in tidy form}
\usage{
tidy(x, ...)
}
\arguments{
\item{x}{phase.test object as returned from \code{phase_test}}

\item{...}{ignored}
}
\description{
\code{tidy.phase.test} expands the tidy method for test.phase objects. It gives a summary of the relevant
parameters and estimates.
}
\examples{
logo <- x3p_read(system.file("csafe-logo.x3p", package="x3ptools"))
print(logo)
}
