% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{x3pheader_to_microns}
\alias{x3pheader_to_microns}
\title{Convert x3p header information to microns from meters}
\usage{
x3pheader_to_microns(x3p)
}
\arguments{
\item{x3p}{x3p data read in using read_x3p or read_bullet}
}
\value{
x3p with header information in microns
}
\description{
Convert x3p header information to microns from meters
}
