% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bullet-scores.R
\name{bullet_to_land_predict}
\alias{bullet_to_land_predict}
\title{Get land to land prediction based on bullet to bullet comparisons}
\usage{
bullet_to_land_predict(land1, land2, scores, difference, alpha = 0.05)
}
\arguments{
\item{land1}{(numeric) vector with land ids of bullet 1}

\item{land2}{(numeric) vector with land ids of bullet 2}

\item{scores}{numeric vector of scores to be summarized into a single number}

\item{difference}{numeric value describing the minimal difference between scores from same source versus different sources.}

\item{alpha}{numeric value describing the significance level for the bootstrap}
}
\value{
numeric vector of binary prediction whether two lands are same-source. Vector has the same length as the input vectors.
}
\description{
The combination of `land1` and `land2` are a key to the scores,
i.e. if a bullet has six lands, each of the input vectors should have
length 36.
}
