% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loess.R
\name{cc_fit_loess}
\alias{cc_fit_loess}
\title{Fit a loess curve to a bullet data frame}
\usage{
cc_fit_loess(ccdata, span = 0.75)
}
\arguments{
\item{ccdata}{The crosscut as returned from x3p_to_df, grooves need to be
removed ahead of time}

\item{span}{The span to use for the loess regression}
}
\value{
a list of a data frame of the original bullet measurements extended
          by loess fit, residuals, and standard errors and two plots: a plot
          of the fit, and a plot of the bullet's land signature.
}
\description{
A loess regression is fit to the surface measurements and residuals are
calculated.
The most extreme 0.25% of residuals are filtered from further consideration.
The result is called the raw signature of the bullet land.
Adapted from `fit_loess` in `bulletr`
}
\examples{
library(dplyr)
ccdata <- data_frame(
  x = seq(0, 6, .001),
  value = 10 - (3 - x)^2 + rnorm(length(x), sd = .25)
)
cc_fit_loess(ccdata = ccdata)
}
