% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grooves.R
\name{cc_locate_grooves}
\alias{cc_locate_grooves}
\title{Find the grooves of a bullet land}
\usage{
cc_locate_grooves(ccdata, method = "rollapply", smoothfactor = 15,
  adjust = 10, groove_cutoff = 400, mean_left = NULL, mean_right = NULL,
  mean_window = 100, return_plot = F, ...)
}
\arguments{
\item{ccdata}{data frame of the crosscut. Data frame needs location x and measured values as `value`. If multiple crosscuts are to be considered, include a variable y and use as a key.}

\item{method}{method to use for identifying grooves. Defaults to "rollapply"}

\item{smoothfactor}{The smoothing window to use - XXX the smoothing window seems to depend on the resolution at which the data has been collected.}

\item{adjust}{positive number to adjust the grooves - XXX should be expressed in microns rather than an index}

\item{groove_cutoff}{The index at which a groove cannot exist past - XXX this parameter should be expressed in microns rather than as an index to be able to properly deal with different resolutions}

\item{mean_left}{If provided, the location of the average left groove}

\item{mean_right}{If provided, the location of the average right groove}

\item{mean_window}{The window around the means to use}

\item{return_plot}{Return plot of grooves?}

\item{...}{parameters passed on to specific groove location methods}
}
\description{
Find the grooves of a bullet land
}
