% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{extract_feature_lag}
\alias{extract_feature_lag}
\title{Extract lag from two (or more) aligned signatures}
\usage{
extract_feature_lag(aligned)
}
\arguments{
\item{aligned}{data frame with variable x (for location) and two or
more measurements}
}
\value{
(vector) of lags
}
\description{
In the case of two signatures the result is an integer definining the number
of indices by which the second signature is shifted compared to the first
signature in the alignment. Note that this lag can be negative.
In the case of multiple signatures the result is a vector of non-negative
integers of the same length as signatures. Each element gives the number of
indices by which the corresponding signature is shifted compared to the
*first* signature. By definition, one of the numbers has to be 0 indicating
the *first* signature.
}
