% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{extract_feature_left_cms}
\alias{extract_feature_left_cms}
\title{Extract number of consecutively matching elevated striation marks from the
left of two aligned signatures}
\usage{
extract_feature_left_cms(striae)
}
\arguments{
\item{striae}{data frame of striation marks based on two aligned signatures}
}
\value{
number of consecutively matching striation marks (from left)
}
\description{
Extract number of consecutively matching elevated striation marks from the
left of two aligned signatures
}
