% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{extract_features_all}
\alias{extract_features_all}
\title{Extract features from aligned signatures}
\usage{
extract_features_all(aligned, striae, ...)
}
\arguments{
\item{aligned}{aligned signatures, result from `sig_cms_max`}

\item{striae}{data frame with evaluated matching striae}

\item{...}{passed on to extractor functions
XXX this needs some fixing}
}
\description{
Extract features from aligned signatures
}
