% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{extract_features_all}
\alias{extract_features_all}
\title{Extract features from aligned signatures}
\usage{
extract_features_all(res)
}
\arguments{
\item{results}{list consisting of data frames of lines and aligned signatures, result from `sig_cms_max`}
}
\description{
Extract features from aligned signatures
}
