% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{extract_features_all}
\alias{extract_features_all}
\title{Extract features from aligned signatures}
\usage{
extract_features_all(aligned, striae, ...)
}
\arguments{
\item{aligned}{aligned signatures, result from `sig_cms_max`}

\item{striae}{data frame with evaluated matching striae}

\item{...}{passed on to extractor functions
TODO: test whether ... is actually passed on correctly}
}
\description{
Extract features from aligned signatures
}
\examples{
\dontrun{
library(bulletxtrctr)
library(x3ptools)

example_data <- bullet_pipeline(
  location = list(
    Bullet1 = c(hamby252demo$bullet1[2]),
    Bullet2 = c(hamby252demo$bullet2[4])
  ),
  x3p_clean = function(x) x \%>\%
      x3pheader_to_microns() \%>\%
      rotate_x3p(angle = -90) \%>\%
      y_flip_x3p()
)

alignment <- sig_align(example_data$sigs[[1]]$sig,
                       example_data$sigs[[2]]$sig)
striae <- sig_cms_max(alignment)

extract_features_all(alignment, striae)
}
}
\seealso{
Other alignment-related-features: \code{\link{extract_feature_D}},
  \code{\link{extract_feature_ccf}},
  \code{\link{extract_feature_lag}},
  \code{\link{extract_feature_length}},
  \code{\link{extract_feature_overlap}}

Other striae-related-features: \code{\link{extract_feature_cms_all}},
  \code{\link{extract_feature_cms_peak_only}},
  \code{\link{extract_feature_left_cms}},
  \code{\link{extract_feature_matches}},
  \code{\link{extract_feature_mismatches}},
  \code{\link{extract_feature_non_cms}},
  \code{\link{extract_feature_right_cms}},
  \code{\link{extract_feature_sum_peaks}},
  \code{\link{extract_helper_feature_n_striae}}
}
