% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phase_test.R
\name{f_T}
\alias{f_T}
\title{Reference density for the test statistic between Same-source and Different-source averages using phase selection}
\usage{
f_T(t, sigma = 1, n = 6)
}
\arguments{
\item{t}{value of the observed test statistic, $t$ is positive.}

\item{sigma}{(estimated) standard deviation of an average of size $n$ of the (original) sample}

\item{n}{integer value, number of values in the same-source group.}
}
\description{
Assuming an overall sample size of 36 (standing for 6 x 6 comparisons, $n=6$), we are interested in the distribution of
T defined as the difference between the average of the same source values and the median average (XXX see paper for definition) of the different source values:
$$T = \bar{X}\emph{{SS} - \tilde{X}}{DS}$$, where groups SS and DS are set by phase selection (after inspecting the set of $n$ values).
For a (one-sided) hypothesis test of $H_0$: no difference between the means and $H_1:$: the same-source values are larger
on average than the different source values, the test statistic $T$ has under the null hypothesis and the group-finding
strategy ${\cal G}$ a distribution given as
$$
P(T < t \mid H_0, {\cal G}) = C \cdot \int_{-\infty}^{\infty} F^2(x) f(x) \left\link[= F(x + t) - F(x) \\right]{ F(x + t) - F(x) \right}^3 dx
$$
where $F$ is the (approximately normal) distribution of $\bar{X}_{SS}$ under $H_0$ with expected value of $\mu$ ($=E\link[=\\bar{X}_{SS}]{\bar{X}_{SS}} = E\link[=\\bar{X}_{DS}]{\bar{X}_{DS}}$) and standard deviation
$\sigma$ (standard deviation of a group of size $n$ under $H_0$).
}
\examples{
# example code
F_T(0.2, sigma = 0.1, lower.tail = FALSE) # only multitude of sigma informs the test
F_T(2, lower.tail = FALSE)
F_T(3, lower.tail = FALSE)
F_T(4, lower.tail = FALSE)
# critical values of t distribution and F_T (for sigma = 1):
F_T(qt(1-0.05, df=34), lower.tail = FALSE)
F_T(qt(1-0.05/6, df=34), lower.tail = FALSE)
F_T(qt(1-0.01/6, df=34), lower.tail = FALSE)

x <- 10^-(1:4)
x <- sort(c(x, x/2))
t_q <- qt(1-x, df=34)
adjusted <- F_T(t_q, lower.tail=FALSE)
}
