% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align.R
\name{get_ccf}
\alias{get_ccf}
\title{Cross correlation function between two vectors}
\usage{
get_ccf(x, y, min.overlap = round(0.1 * max(length(x), length(y))))
}
\arguments{
\item{x}{vector, assumption is that x is longer than y}

\item{y}{vector}

\item{min.overlap}{integer value: what is the minimal number of values
between x and y that should be considered?}
}
\value{
list with ccf values and lags
}
\description{
Calculate the lagged correlation between numeric vectors x and y.
Vectors x and y are assumed to be captured at the same resolution and,
similarly,
successive values in x and y are assumed to be equi-distant.
Missing values are allowed in each vector, correlations are calculated
based on the complete cases.
}
\details{
This version of the cross correlation function is different from the \code{stats}
implementation of \code{ccf} in two ways:
\enumerate{
\item We consider the full region of correlations between vectors x and y as
specified by min.overlap rather than just the overlap. The two vectors can
be of very different
length (e.g. when y is just a snippet recovered from a crime scene and x is
from the full length object in the lab).
\item We do not use a Fourier transformation to calculate cross-correlation.
This makes the evaluation slower, but prevents any edge effects.
}
}
\examples{
library(dplyr)
x <- runif(20)
get_ccf(x, lead(x, 5))
get_ccf(x, lag(x, 5), min.overlap = 3)
x <- runif(100)
get_ccf(x[45:50], x, min.overlap = 6)
}
