% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align.R
\name{get_ccf}
\alias{get_ccf}
\title{Cross correlation function between two vectors}
\usage{
get_ccf(x, y, min.overlap = 0.1 * max(length(x), length(y)))
}
\arguments{
\item{x}{vector}

\item{y}{vector}

\item{min.overlap}{integer value: what is the minimal number of values between x and y that should be considered?}
}
\value{
list with ccf values and lags
}
\description{
Cross correlation function between two vectors
}
\examples{
library(dplyr)
x <- runif(20)
get_ccf(x, lead(x, 5))
get_ccf(x, lag(x, 5), min.overlap=3)
x <- runif(100)
get_max_ccf(x[45:50], x, min.overlap=6)
}
