% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth.R
\name{raw_sig_smooth}
\alias{raw_sig_smooth}
\title{Smooth the raw signature}
\usage{
raw_sig_smooth(sig, span = 0.03, limits = c(-5, 5))
}
\arguments{
\item{sig}{numeric vector of the raw signature}

\item{span}{width of the smoother, defaults to 0.03. XXX Should be}

\item{limits}{vector of the form c(min, max). Results will be limited to be
between these values.}
}
\value{
numeric vector of the same length as the input with the smoothed
          signature.
}
\description{
Smooth the raw signature
}
\examples{
\dontrun{
library(ggplot2)
x <- seq(0, 6 * pi, by = .02)
y <- 5.5 * sin(x) + rnorm(length(x))


ggplot() +
  geom_point(aes(x = 1:length(y), y = y), shape = 1) +
  geom_line(aes(x = 1:length(y), y = smoothloess(y, .05)),
    color = "red"
  ) +
  geom_line(aes(x = 1:length(y), y = raw_sig_smooth(y, .05)),
    color = "blue"
  )
}
}
