% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runExample.R
\name{runExample}
\alias{runExample}
\title{Execute the Shiny App for bullet investigation}
\usage{
runExample()
}
\description{
This shiny app would expect a tibble object called `shiny.tt` in the user's working environment.
}
\details{
shiny.tt should have variables: `x3p`, `scan_id`, `crosscut`, `ccdata`, `grooves`.

To have this shiny app run properly, `shiny.tt` should at least include `grooves` as a variable
}
\examples{
\dontrun{
bulletxtrctr::runExample()
}
}
