% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cms.R
\name{sig_cms_max}
\alias{sig_cms_max}
\title{Identify the number of maximum CMS between two signatures}
\usage{
sig_cms_max(sig1, sig2, column = "resid", span = 35)
}
\arguments{
\item{sig1}{vector of smoothed first signature}

\item{sig2}{vector of smoothed second signature}

\item{column}{The column which to consider for the signature}

\item{span}{positive number  for the smoothfactor to use for assessing peaks.}
}
\value{
list of matching parameters, data set of the identified striae, and the aligned data sets.
}
\description{
adapted from `bulletGetMaxCMS`
aligns two signatures, identifies peaks and valleys, matches striae, and counts longest run.
Should probably be split into four or five sub functions.
}
