% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peaks.R
\name{sig_get_peaks}
\alias{sig_get_peaks}
\title{Identify the location and the depth of peaks and valleys in a signature}
\usage{
sig_get_peaks(sig, smoothfactor = 35, striae = TRUE, window = TRUE)
}
\arguments{
\item{sig}{numeric vector of smoothed signature, padded with NAs at front
and/or back.}

\item{smoothfactor}{set to default of 35. Smaller values will pick up on
smaller changes in the crosscut.}

\item{striae}{If TRUE, show the detected striae on the plot}

\item{window}{If TRUE, show the window of the striae on the plot}
}
\value{
list of several objects:
}
\description{
Adapted from the function `get_peaks`
}
\examples{
\dontrun{
# Set the data up to be read in, cleaned, etc.
library(bulletxtrctr)
library(x3ptools)
library(ggplot2)

example_data <- bullet_pipeline(
  location = list(Bullet1 = c(hamby252demo$bullet1[3])),
  x3p_clean = function(x) x \%>\%
      x3pheader_to_microns() \%>\%
      rotate_x3p(angle = -90) \%>\%
      y_flip_x3p()
)

sig <- cc_get_signature(example_data$ccdata[[1]], example_data$grooves[[1]])
sig_peaks <- sig_get_peaks(sig$sig)
sig_peaks[1:6]
# Plot
sig_peaks$plot
}
}
