% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crosscut.R
\name{x3p_crosscut}
\alias{x3p_crosscut}
\title{Read a crosscut from a 3d surface file}
\usage{
x3p_crosscut(x3p, y = NULL)
}
\arguments{
\item{x3p}{if character, path to an x3p file. Otherwise a scan in x3p
format is expected.}

\item{y}{level of the crosscut to be taken. If this level does not exist,
the crosscut along the middle of the land is returned.}
}
\value{
data frame
}
\description{
Read a crosscut from a 3d surface file
}
